SVC: hide details in pimpl
